Add files via upload
Fixed two errors , in first error user and admin was able to login through each others tokens. in second i handled the routing of profile,register and login page, i.e if the user is login the register and login pages will be hidden, if the user is not login the profile button will be hidden.